fix cast
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@40614 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/src/jnetcap.c
@@ -554,7 +554,7 @@ static void              _conntrack_hook( int type, long mark, long conntrack_id
     /* Call the global method */
     (*env)->CallVoidMethod( env, global_hook, _jnetcap.java.conntrack_event_method_id,
                             ((jlong)session_id),
-                            ((jint)type), ((jint)mark), ((jlong)conntrack_id),
+                            ((jint)type), ((jlong)mark), ((jlong)conntrack_id),
                             ((jint)l3_proto), ((jint)l4_proto),
                             ((jlong)c_client_addr), ((jlong)c_server_addr),
                             ((jint)c_client_port), ((jint)c_server_port),
@@ -632,7 +632,7 @@ static void              _hook( int protocol, netcap_session_t* netcap_sess, voi
         session_id = netcap_sess->session_id;
         
         /* Call the global method */
-        (*env)->CallVoidMethod( env, global_hook, _jnetcap.java.event_method_id, netcap_sess->session_id );
+        (*env)->CallVoidMethod( env, global_hook, _jnetcap.java.event_method_id, ((jlong)netcap_sess->session_id) );
 
         debug( 10, "JNETCAP: Exiting hook\n" );
